+++ packages/commercetools/composables/__tests__/useCheckout/useCheckout.spec.ts
@@ -107,6 +107,7 @@ describe('[commercetools-composables] useCheckout/setShippingDetails', () => {
     await setShippingDetails(shippingAddress);
 
     expect(shippingDetails.value).toEqual({ ...shippingAddress, contactInfo: {} });
+    expect(loading.value.shippingAddress).toBe(false);
   });
 
   it('overwrites contactInfo properly', async () => {
@@ -116,6 +117,7 @@ describe('[commercetools-composables] useCheckout/setShippingDetails', () => {
     await setShippingDetails(shippingAddress);
 
     expect(shippingDetails.value).toEqual({ ...shippingAddress, contactInfo: { phone: '123456789', email: 'test@test.com' } });
+    expect(loading.value.shippingAddress).toBe(false);
   });
 
   it('send shipping details to the api', async () => {
@@ -133,6 +135,31 @@ describe('[commercetools-composables] useCheckout/setShippingDetails', () => {
       version: 1
     }, undefined);
     expect(initFields).toHaveBeenCalledWith({ id: 'new-cart-id' });
+    expect(loading.value.shippingAddress).toBe(false);
+  });
+
+  it('fails when updateCart throws', async () => {
+    (updateCart as jest.Mock).mockImplementationOnce(() => {
+      throw 'error';
+    });
+    const setShippingDetails = createSetShippingDetails({ factoryParams: {}, cartFields, setCart });
+    const shippingAddress = { firstName: 'John', lastName: 'Doe' };
+    try {
+      await setShippingDetails(shippingAddress, { save: true });
+      fail();
+    } catch {
+      expect(shippingDetails.value).toEqual({ ...shippingAddress, contactInfo: {} });
+      expect(updateCart).toHaveBeenCalledWith({
+        actions: [
+          'setShippingMethodAction',
+          'setShippingAddressAction'
+        ],
+        id: 'cart-id',
+        version: 1
+      }, undefined);
+      expect(initFields).not.toBeCalled();
+      expect(loading.value.shippingAddress).toBe(false);
+    }
   });
 });
 
@@ -148,6 +175,7 @@ describe('[commercetools-composables] useCheckout/setBillingDetails', () => {
     await setBillingDetails(billingAddress);
 
     expect(billingDetails.value).toEqual({ ...billingAddress, contactInfo: {} });
+    expect(loading.value.billingAddress).toBe(false);
   });
 
   it('send billing details to the API', async () => {
@@ -164,6 +192,30 @@ describe('[commercetools-composables] useCheckout/setBillingDetails', () => {
       version: 1
     }, undefined);
     expect(initFields).toHaveBeenCalledWith({ id: 'new-cart-id' });
+    expect(loading.value.billingAddress).toBe(false);
+  });
+
+  it('fails when updateCart throws', async () => {
+    (updateCart as jest.Mock).mockImplementationOnce(() => {
+      throw 'error';
+    });
+    const setBillingDetails = createSetBillingDetails({ factoryParams: {}, cartFields, setCart });
+    const billingAddress = { firstName: 'John', lastName: 'Doe' };
+    try {
+      await setBillingDetails(billingAddress, { save: true });
+      fail();
+    } catch {
+      expect(billingDetails.value).toEqual({ ...billingAddress, contactInfo: {} });
+      expect(updateCart).toHaveBeenCalledWith({
+        actions: [
+          'setBillingAddressAction'
+        ],
+        id: 'cart-id',
+        version: 1
+      }, undefined);
+      expect(initFields).not.toBeCalled();
+      expect(loading.value.billingAddress).toBe(false);
+    }
   });
 });
 
@@ -208,6 +260,7 @@ describe('[commercetools-composables] useCheckout/loadShippingMethods', () => {
 
     expect(getShippingMethods).not.toBeCalled();
     expect(setShippingMethod).not.toBeCalled();
+    expect(loading.value.shippingMethods).toBe(false);
   });
 
   it('loads shipping methods with empty list', async () => {
@@ -220,6 +273,7 @@ describe('[commercetools-composables] useCheckout/loadShippingMethods', () => {
     expect(shippingMethods.value).toEqual([]);
     expect(chosenShippingMethod.value).toEqual({});
     expect(setShippingMethod).not.toBeCalled();
+    expect(loading.value.shippingMethods).toBe(false);
   });
 
   it('selects shipping method from the cart', async () => {
@@ -236,6 +290,29 @@ describe('[commercetools-composables] useCheckout/loadShippingMethods', () => {
     expect(shippingMethods.value).toEqual([]);
     expect(chosenShippingMethod.value).toEqual({ });
     expect(setShippingMethod).not.toBeCalled();
+    expect(loading.value.shippingMethods).toBe(false);
+  });
+
+  it('fails when getShippingMethods throws', async () => {
+    (getShippingMethods as any).mockImplementationOnce(() => {
+      throw 'error';
+    });
+    cart.value = {
+      ...cart.value,
+      shippingInfo: { shippingMethod: { method: 'cart' } }
+    } as any;
+    const setShippingMethod = jest.fn();
+    const loadShippingMethods = createLoadShippingMethods({ factoryParams: {}, setShippingMethod, cartFields });
+    try {
+      await loadShippingMethods();
+      fail();
+    } catch {
+      expect(getShippingMethods).toBeCalled();
+      expect(shippingMethods.value).toEqual([]);
+      expect(chosenShippingMethod.value).toEqual({ });
+      expect(setShippingMethod).not.toBeCalled();
+      expect(loading.value.shippingMethods).toBe(false);
+    }
   });
 });
 
@@ -250,6 +327,21 @@ describe('[commercetools-composables] useCheckout/placeOrder', () => {
     await placeOrder();
 
     expect(createMyOrderFromCart).toBeCalled();
+    expect(loading.value.order).toBe(false);
+  });
+
+  it('fails when createMyOrderFromCart throws', async () => {
+    (createMyOrderFromCart as jest.Mock).mockImplementationOnce(() => {
+      throw 'error';
+    });
+    const placeOrder = createPlaceOrder({ cartFields });
+    try {
+      await placeOrder();
+      fail();
+    } catch {
+      expect(createMyOrderFromCart).toBeCalled();
+      expect(loading.value.order).toBe(false);
+    }
   });
 });
 
@@ -284,6 +376,7 @@ describe('[commercetools-composables] useCheckout/setPersonalDetails', () => {
     expect(personalDetails.value).toEqual(data);
     expect(updateCart).not.toBeCalled();
     expect(setShippingDetails).not.toBeCalled();
+    expect(loading.value.personalDetails).toBe(false);
   });
 
   it('set personal details and send it to the API', async () => {
@@ -300,6 +393,30 @@ describe('[commercetools-composables] useCheckout/setPersonalDetails', () => {
     }, undefined);
     expect(setShippingDetails).toBeCalled();
     expect(initFields).toBeCalled();
+    expect(loading.value.personalDetails).toBe(false);
+  });
+
+  it('fails when updateCart throws', async () => {
+    (updateCart as jest.Mock).mockImplementationOnce(() => {
+      throw 'error';
+    });
+    const setShippingDetails = jest.fn();
+    const setPersonalDetails = createSetPersonalDetails({ factoryParams: {}, setShippingDetails, cartFields, setCart });
+    const data = { firstName: 'John', lastName: 'Doe' };
+    try {
+      await setPersonalDetails(data, { save: true });
+      fail();
+    } catch {
+      expect(personalDetails.value).toEqual(data);
+      expect(updateCart).toHaveBeenCalledWith({
+        actions: ['setCustomerEmail'],
+        id: 'cart-id',
+        version: 1
+      }, undefined);
+      expect(setShippingDetails).not.toBeCalled();
+      expect(initFields).not.toBeCalled();
+      expect(loading.value.personalDetails).toBe(false);
+    }
   });
 });
 
@@ -316,6 +433,7 @@ describe('[commercetools-composables] useCheckout/setShippingMethod', () => {
 
     expect(chosenShippingMethod.value).toEqual(method);
     expect(updateCart).not.toBeCalled();
+    expect(loading.value.shippingMethod).toBe(false);
   });
 
   it('set shipping method and send it to the API', async () => {
@@ -330,6 +448,28 @@ describe('[commercetools-composables] useCheckout/setShippingMethod', () => {
       version: 1
     }, undefined);
     expect(initFields).toBeCalled();
+    expect(loading.value.shippingMethod).toBe(false);
+  });
+
+  it('fails when updateCart throws', async () => {
+    (updateCart as jest.Mock).mockImplementationOnce(() => {
+      throw 'error';
+    });
+    const setShippingMethod = createSetShippingMethod({ factoryParams: {}, cartFields, setCart });
+    const method = { name: 'method1', id: 1};
+    try {
+      await setShippingMethod(method, { save: true });
+      fail();
+    } catch {
+      expect(chosenShippingMethod.value).toEqual(method);
+      expect(updateCart).toHaveBeenCalledWith({
+        actions: ['setShippingMethodAction'],
+        id: 'cart-id',
+        version: 1
+      }, undefined);
+      expect(initFields).not.toBeCalled();
+      expect(loading.value.shippingMethod).toBe(false);
+    }
   });
 });